apps/accf/requirements.txt
# Core Framework Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database and Graph Dependencies
neo4j>=5.15.0
neo4j-graphrag[openai]>=1.0.0

# OpenAI Integration
openai>=1.12.0

# Async and Concurrency
asyncio

# Caching and Session Management
redis>=5.0.0

# Environment and Configuration
python-dotenv>=1.0.0

# Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=6.2.1

# Code Quality and Development Tools
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
mutmut>=3.3.1
safety>=3.6.0

# Data Processing and Serialization
pyyaml>=6.0.0

# MCP (Model Context Protocol) Support
mcp>=1.0.0

# Type Extensions
typing-extensions>=4.8.0

# Logging and Monitoring
structlog>=23.2.0

# Security and Validation
cryptography>=44.0.1  # Fixed CVE-2024-12797

# HTTP Client
httpx>=0.25.0

# Date and Time Handling
python-dateutil>=2.8.0

# Web Framework (if needed for compatibility)
flask>=3.1.1  # Fixed CVE-2025-47278

# WSGI Server (if needed for production)
gunicorn>=23.0.0  # Fixed CVE-2024-6827, CVE-2024-1135

# JWT and Security (commented out - not currently used)
# python-jose[cryptography]>=3.3.2  # Fixed CVE-2024-33664, CVE-2024-33663

# File Upload Handling
python-multipart>=0.0.18  # Fixed CVE-2024-53981

# Cryptography (commented out - not currently used)
# ecdsa>=0.19.1  # Latest available version

# AWS Integration (for secrets management)
boto3>=1.39.1

# Additional dependencies found in current environment
aiofiles>=24.1.0
aiohttp>=3.12.14
anyio>=4.9.0
attrs>=25.3.0
click>=8.1.8
certifi>=2025.6.15
